+++ test/src/org/omegat/filters/HTMLFilter2Test.java
@@ -34,12 +34,13 @@ import java.util.List;
 import java.util.Map;
 
 import org.junit.Test;
+
 import org.omegat.core.Core;
 import org.omegat.core.data.IProject;
 import org.omegat.filters2.FilterContext;
 import org.omegat.filters2.html2.HTMLFilter2;
 import org.omegat.filters2.html2.HTMLOptions;
-import org.omegat.filters2.html2.HTMLUtils;
+import org.omegat.util.HTMLUtils;
 import org.omegat.util.Language;
 import org.omegat.util.OStrings;
