+++ test/src/org/omegat/util/HTMLUtilsTest.java
@@ -23,7 +23,7 @@
  You should have received a copy of the GNU General Public License
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  **************************************************************************/
-package org.omegat.filters2.html2;
+package org.omegat.util;
 
 import org.junit.Test;
